pyproject.toml
[project]
name = "easy-agent"
version = "0.1.0"
description = "A white-box, extensible agent engineering foundation for Python."
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "MIT" }
authors = [{ name = "Codex" }]
dependencies = [
  "anyio>=4.11.0",
  "httpx>=0.28.1",
  "httpx-sse>=0.4.3",
  "mcp>=1.21.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.11.0",
  "pyyaml>=6.0.2",
  "rich>=14.1.0",
  "tenacity>=9.1.2",
  "typer>=0.20.0",
]

[dependency-groups]
dev = [
  "mypy>=1.18.2",
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "respx>=0.22.0",
  "ruff>=0.13.3",
  "types-PyYAML>=6.0.12.20250915",
]

[project.scripts]
easy-agent = "easy_agent.cli:app"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/easy_agent"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]
ignore = ["E501", "ASYNC220", "ASYNC240"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-p", "no:cacheprovider"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
packages = ["easy_agent"]

[[tool.mypy.overrides]]
module = ["yaml"]
ignore_missing_imports = true
